telegram-news-analysis/src/collector/requirements.txt
telethon==1.39.0
pymongo==4.12.0
python-dotenv==0.20.0
motor==3.7.0
asyncpg==0.30.0
